+++ lib/model/settings/settings.dart
@@ -29,6 +29,8 @@ class Settings extends ChangeNotifier {
 
   Settings._private();
 
+  static const int moveDestinationAlbumMax = 3;
+
   static const Set<String> _internalKeys = {
     hasAcceptedTermsKey,
     catalogTimeZoneKey,
@@ -37,6 +39,7 @@ class Settings extends ChangeNotifier {
     platformAccelerometerRotationKey,
     platformTransitionAnimationScaleKey,
     topEntryIdsKey,
+    moveDestinationAlbumsKey,
   };
   static const _widgetKeyPrefix = 'widget_';
 
@@ -51,6 +54,7 @@ class Settings extends ChangeNotifier {
   static const tileLayoutPrefixKey = 'tile_layout_';
   static const entryRenamingPatternKey = 'entry_renaming_pattern';
   static const topEntryIdsKey = 'top_entry_ids';
+  static const moveDestinationAlbumsKey = 'move_destination_albums';
 
   // display
   static const displayRefreshRateModeKey = 'display_refresh_rate_mode';
@@ -313,6 +317,10 @@ class Settings extends ChangeNotifier {
   List<int>? get topEntryIds => getStringList(topEntryIdsKey)?.map(int.tryParse).whereNotNull().toList();
 
   set topEntryIds(List<int>? newValue) => setAndNotify(topEntryIdsKey, newValue?.map((id) => id.toString()).whereNotNull().toList());
+
+  List<String> get moveDestinationAlbums => getStringList(moveDestinationAlbumsKey) ?? [];
+
+  set moveDestinationAlbums(List<String> newValue) => setAndNotify(moveDestinationAlbumsKey, newValue.take(Settings.moveDestinationAlbumMax).toList());
 
   // display